Update v0.3
A lot of changes. Too many to count. However, this time I have made a proper changelog file! If it properly includes into the repository, I will leave it at that. But if not, I will post a link to a either a dropbox or a seperate github repository for it for those who are interested.